refactor(app): split generate gallery
